WIP PBR, preparing for blending, and bugfixes

+++ Src/Assets/AssetLoader.h
@@ -35,9 +35,8 @@ private:
 public:
 	std::shared_ptr<SLoadedGLTF> LoadGLTFScene(const std::filesystem::path& filePath);
 
-	// TODO: these return SGlTextureId
 	static std::optional<SGlTextureId> LoadTexture2DFromFile(std::filesystem::path const& texturePath, bool bFlipVertical);
-	static std::optional<SGlTextureId> LoadTexture2DFromBuffer(void* buffer, int size, bool bFlipVertical);
+	static std::optional<SGlTextureId> LoadTexture2DFromBuffer(void* buffer, size_t size, bool bFlipVertical);
 
 	static std::optional<CGlShader> LoadShaderProgram(const std::filesystem::path& vsPath, const std::filesystem::path& fsPath);
